+++ frontend/src/components/study/PersonalStudySidebar.module.scss
@@ -116,94 +116,7 @@
 
 
 
-  // 검색 영역
-  .search-area {
-    padding: 0 16px 16px 16px;
-    background-color: var(--background-color);
-    display: flex;
-    gap: 10px;
-    align-items: flex-start;
 
-    .search-input-wrapper {
-      position: relative;
-      flex: 1;
-
-      .search-input {
-        width: 100%;
-        background-color: var(--white-color);
-        border: 1px solid var(--secondary-color);
-        border-radius: 12px;
-        padding: 12px 40px 12px 12px;
-        font-size: 14px;
-        box-sizing: border-box;
-        transition: all 0.2s;
-        letter-spacing: -0.2px;
-
-        &:focus {
-          outline: none;
-          border-color: var(--text-color);
-          box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
-        }
-
-        &::placeholder {
-          color: #999;
-        }
-      }
-
-      .clear-button {
-        position: absolute;
-        right: 8px;
-        top: 50%;
-        transform: translateY(-50%);
-        background: none;
-        border: none;
-        color: #999;
-        font-size: 16px;
-        cursor: pointer;
-        padding: 4px;
-        border-radius: 50%;
-        width: 24px;
-        height: 24px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        transition: all 0.2s;
-
-        &:hover {
-          background-color: rgba(0, 0, 0, 0.05);
-          color: var(--text-color);
-        }
-
-        &:active {
-          transform: translateY(-50%) scale(0.9);
-        }
-      }
-    }
-
-    .search-button {
-      background-color: var(--secondary-color);
-      color: var(--white-color);
-      border: none;
-      border-radius: 12px;
-      padding: 12px 20px;
-      font-size: 14px;
-      font-weight: 600;
-      cursor: pointer;
-      box-shadow: 0 2px 8px rgba(144, 202, 249, 0.3);
-      transition: all 0.2s;
-      letter-spacing: -0.2px;
-      white-space: nowrap;
-
-      &:hover {
-        transform: translateY(-1px);
-        box-shadow: 0 4px 12px rgba(144, 202, 249, 0.4);
-      }
-
-      &:active {
-        transform: scale(0.98);
-      }
-    }
-  }
 
   // 단어 목록
   .word-list {
